cccp: cast body size to unsigned long before printing
Reviewed-on: http://review.couchbase.org/34020
Tested-by: Mark Nunberg <mnunberg@haskalah.org>

+++ src/bucketconfig/bc_cccp.c
@@ -419,7 +419,7 @@ static void io_read_handler(lcb_connection_t conn)
                 "CCCP Packet responded with 0x%x; nkey=%d, nbytes=%lu, cmd=0x%x, seq=0x%x",
                 PACKET_STATUS(&pi),
                 PACKET_NKEY(&pi),
-                PACKET_NBODY(&pi),
+                (unsigned long)PACKET_NBODY(&pi),
                 PACKET_OPCODE(&pi),
                 PACKET_OPAQUE(&pi));
